currentViewersをmessagesViewから分離
