+++ src/components/Main/MainView/ChannelView/ChannelViewContent/ChannelViewContent.vue
@@ -13,13 +13,14 @@
       :channel-id="channelId"
       :entry-message-id="entryMessageId"
       :pinned-messages="pinnedMessages"
+      :typing-users="typingUsers"
     />
   </div>
 </template>
 
 <script lang="ts">
 import { computed, ref, Ref, toRef } from 'vue'
-import { ChannelId } from '/@/types/entity-ids'
+import { ChannelId, UserId } from '/@/types/entity-ids'
 import { debounce, throttle } from 'throttle-debounce'
 import { useMessageInputStateAttachment } from '/@/store/ui/messageInputState'
 import { useToastStore } from '/@/store/ui/toast'
@@ -81,6 +82,7 @@ const props = defineProps<{
   channelId: ChannelId
   entryMessageId?: ChannelId
   pinnedMessages: Pin[]
+  typingUsers: UserId[]
 }>()
 
 const { canDrop, onDrop, onDragStart, onDragOver } = useDragDrop(
