+++ src/store/domain/messagesView.ts
@@ -1,15 +1,14 @@
-import { ChannelViewer, ChannelViewState, Message } from '@traptitech/traq'
+import { ChannelViewState, Message } from '@traptitech/traq'
 import { EmbeddingOrUrl, ExternalUrl } from '@traptitech/traq-markdown-it'
 import { defineStore, acceptHMRUpdate } from 'pinia'
-import { computed, ref } from 'vue'
+import { ref } from 'vue'
 import apis from '/@/lib/apis'
 import { isExternalUrl, isFile, isMessage } from '/@/lib/guard/embeddingOrUrl'
 import { render } from '/@/lib/markdown/markdown'
-import { changeViewState, wsListener } from '/@/lib/websocket'
+import { changeViewState } from '/@/lib/websocket'
 import { convertToRefsStore } from '/@/store/utils/convertToRefsStore'
 import { ChannelId, ClipFolderId, MessageId } from '/@/types/entity-ids'
 import { messageMitt, useMessagesStore } from '/@/store/entities/messages'
-import { useMeStore } from '/@/store/domain/me'
 
 export type LoadingDirection = 'former' | 'latter' | 'around' | 'latest'
 interface BaseGetMessagesParams {
@@ -61,7 +60,6 @@ const isIncludedHost = (url: ExternalUrl) => {
 
 // FIXME: 分離
 const useMessagesViewPinia = defineStore('domain/messagesView', () => {
-  const meStore = useMeStore()
   const messagesStore = useMessagesStore()
 
   /** 現在のチャンネルID、日時ベースのフェッチを行う */
@@ -76,38 +74,13 @@ const useMessagesViewPinia = defineStore('domain/messagesView', () => {
   const receiveLatestMessages = ref(false)
   const renderedContentMap = ref(new Map<MessageId, string>())
   const embeddingsMap = ref(new Map<MessageId, EmbeddingOrUrl[]>())
-  /** チャンネルを見ている人の一覧(古い順) */
-  const currentViewers = ref<ChannelViewer[]>([])
   /** 現在編集中のメッセージID */
   const editingMessageId = ref<MessageId>()
-
-  /**
-   * チャンネルを見ている人(入力中も含む)のIDの一覧(古い順)
-   */
-  const viewingUsers = computed(() =>
-    currentViewers.value
-      .filter(
-        v => v.state === ChannelViewState.Monitoring || ChannelViewState.Editing
-      )
-      .map(v => v.userId)
-  )
-
-  /**
-   * チャンネルで入力中の人のIDの一覧(新しい順)
-   */
-  const typingUsers = computed(() => {
-    const myId = meStore.myId.value
-    return currentViewers.value
-      .filter(v => v.state === ChannelViewState.Editing && v.userId !== myId)
-      .map(v => v.userId)
-      .reverse()
-  })
 
   const resetViewState = () => {
     currentChannelId.value = undefined
     currentClipFolderId.value = undefined
     renderedContentMap.value = new Map()
-    currentViewers.value = []
   }
 
   const changeCurrentChannel = (payload: {
@@ -229,11 +202,7 @@ const useMessagesViewPinia = defineStore('domain/messagesView', () => {
     }
   }
 
-  wsListener.on('CHANNEL_VIEWERS_CHANGED', ({ viewers }) => {
-    currentViewers.value = viewers
-  })
   // 再接続時の再取得はmessagesFetcherで行う
-
   messageMitt.on('updateMessage', async message => {
     if (currentChannelId.value !== message.channelId) return
     await updateAndRenderMessageId(message)
@@ -246,8 +215,6 @@ const useMessagesViewPinia = defineStore('domain/messagesView', () => {
     renderedContentMap,
     embeddingsMap,
     receiveLatestMessages,
-    viewingUsers,
-    typingUsers,
     fetchMessagesByChannelId,
     fetchMessagesInClipFolder,
     renderMessageContent,
